pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-state-compiler"
version = "0.1.0"
description = "Transform fragmented clinical information into a single, safe source of truth"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ansh Shetty"}
]
keywords = ["medgemma", "clinical", "healthcare", "AI", "NLP"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.44.0",
    "accelerate>=0.27.0",
    "huggingface_hub>=0.23.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
]
gpu = [
    "bitsandbytes>=0.42.0",
]

[project.scripts]
csc = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
